requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
tqdm
beautifulsoup4